requirements.txt
openai
python-dotenv
deeplake
langchain
tiktoken
# newspaper3k
streamlit